Make MoveFlags into enum class and update Zobrist hashing for castling status

+++ src/move.cpp
@@ -10,8 +10,8 @@ std::string Move::to_string() const {
     to_return.push_back(this->get_dest_file() + 97);
     to_return.push_back(this->get_dest_rank() + 49);
 
-    if (this->get_move_flags() >= 8) {
-        switch (this->get_move_flags() & 0b0011) {
+    if (static_cast<int>(this->get_move_flags()) >= 8) {
+        switch (static_cast<int>(this->get_move_flags()) & 0b0011) {
         case 0:
             to_return.push_back('r');
             break;
